feat: 관리자 화면 상품 옵션 추가 기능 추가

+++ src/main/java/gift/product/application/ProductService.java
@@ -53,10 +53,11 @@ public class ProductService {
     public ProductResponse createProduct(ProductRequest request) {
         Category category = categoryRepository.findByName(request.categoryName())
                 .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
+        Product product = productRepository.save(ProductMapper.toEntity(request, category));
 
-        return ProductMapper.toResponseDto(
-                productRepository.save(ProductMapper.toEntity(request, category))
-        );
+        optionRepository.save(OptionMapper.toEntity(request.option(), product));
+
+        return ProductMapper.toResponseDto(product);
     }
 
     @Transactional
@@ -75,6 +76,13 @@ public class ProductService {
 
     @Transactional
     public void deleteOptionFromProduct(Long id, OptionRequest request) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND));
+        if (product.getOptions()
+                   .size() == 1) {
+            throw new CustomException(ErrorCode.OPTION_REMOVE_FAILED);
+        }
+
         optionRepository.deleteByProduct_IdAndName(id, request.name());
     }
